.history/requirements_20250602183829.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
blinker==1.9.0
branca==0.8.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.0
coolsms_python_sdk==2.0.3
cryptography==44.0.3
cycler==0.12.1
daphne==4.1.2
debugpy==1.8.5
decorator==5.1.1
display==1.0.0
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
emoji==2.14.1
et_xmlfile==2.0.0
executing==2.1.0
filelock==3.18.0
firebase-admin==6.4.0
folium==0.18.0
fonttools==4.54.1
frozenlist==1.6.0
fsspec==2025.5.0
geojson==2.5.0
geokakao==0.1.1
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googletrans==4.0.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.4
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
jpype1==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.2
jupyter_core==5.7.2
JustWatch @ git+https://github.com/dawoudt/JustWatchAPI.git@e2d8563b936e2ddeee73a930e17d20adc5971686
kaleido==0.2.1
keybert==0.9.0
kiwisolver==1.4.7
knnimpute==0.1.0
konlpy==0.6.0
koreanize-matplotlib==0.1.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mlxtend==0.23.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
narwhals==1.37.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.2
openai==0.28.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.2
plotly==5.24.1
prompt_toolkit==3.0.47
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==6.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyogrio==0.10.0
pyOpenSSL==25.0.0
pyowm==3.3.0
pyparsing==3.2.0
pyproj==3.7.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.13.0
redis==6.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==4.1.0
service-identity==24.2.0
setuptools==80.8.0
shapely==2.0.6
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
solapi==5.0.1
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.45.0
sympy==1.14.0
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
wordcloud==1.9.4
xyzservices==2024.9.0
yarl==1.20.0
yt-dlp==2025.5.22
zodiac-sign==0.2.5
zope.interface==7.2
